# Trivial C optimizations: eliminating unnecessary double promotion, division, and modulo operations