Tested in conjunction with https://github.com/jakubcerveny/gilbert/pull/13.

Even with gcc's `-O3` flag from the Makefile, I was able to see the following savings on a 2020 M1 MacBook Air.  These results omitted the printf per step to only measure the actual algorithm computation and are averaged over three executions each.

* ./gilbert d2xy 490 490
  * before:  49.86 ms
  * after:  37.34 ms

* ./gilbert d2xyz 490 490 490
  * before:  21.19 s
  * after:  17.52 s